pyproject.toml
[project]
name = "ai-research-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "lightning>=2.5.2",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.0",
    "torch>=2.7.1",
    "types-pyyaml>=6.0.12.20250516",
]

[tool.mypy]
ignore_missing_imports = true
